blooddonateproject/requirements.txt
asgiref==3.7.2
cachetools==5.3.3
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google-auth==2.28.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pytz==2024.1
rsa==4.9
sqlparse==0.4.4
